Added more colors!
Learned what an internal stylesheet is and made the bear site practically illegible with goofy ass colors! I tried to figure out how to change the background to make them less of an eyesore but that seems like knowledge that'll come later. VS Code has an insane number of built-in colors to choose from though. 10/10 best editor.